requirements.txt
streamlit>=1.32.0
langchain-community==0.0.29
pypdf==3.17.0
openai>=1.12.0
python-dotenv>=1.0.0
faiss-cpu==1.7.4
tiktoken==0.6.0